+++ package/package.json
@@ -76,7 +76,7 @@
     "react-native": "0.66.2",
     "react-native-builder-bob": "^0.18.2",
     "ts-jest": "^27.0.7",
-    "typescript": "^4.6.2"
+    "typescript": "^4.6.4"
   },
   "dependencies": {
     "react-reconciler": "^0.26.2"
